agent_listener/requirements.txt
web3==7.4.0
python-dotenv==1.0.1
aiohttp==3.9.0
